code/backend/pyproject.toml
[tool.poetry]
name = "drone-imagery-backend"
version = "0.1.0"
description = "Backend API for drone imagery processing with Node ODM"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
httpx = "^0.25.2"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
requests = "^2.31.0"
pyodm = "^1.5.12"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:main"
